Bare bones Android app
Currently consists of a button, which onclick executes an async http request to the API to get the articles metadata.  Returned JSON is mapped to Java objects using Gson.  This as much more work than Python's json.loads(), oh well, when in Rome...  Note that the final object is printed to the console, there is no update of the app view - yet.